app/function_ingest/requirements.txt
pandas==2.2.2
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
